// src/app/Interfaces/View/get-all-visits.ts
export interface GetAllVisits {

    visitsNo?: number;
    customerCode?: number;
    customerName?: string;
    customerLatName?: string;
    salesRepId?: number;
    salesRepName?: string;
    salesRepLatName?: string;
    regionName?: string;
    sectorName?: string;
    territoryName?: string;
    governorateName?: string;
    cclName?: string;
    adress?: string;
    companyName?: string;

}